# services/backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.UpdateUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Channel:
    properties:
      id:
        type: integer
      link:
        type: string
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.News:
    properties:
      body:
        type: string
      channel:
        $ref: '#/definitions/models.Channel'
      channelID:
        type: integer
      createdAt:
        type: string
      duplicates:
        items:
          $ref: '#/definitions/models.News'
        type: array
      id:
        type: integer
      messageLink:
        type: string
      stopThemes:
        items:
          $ref: '#/definitions/models.StopTheme'
        type: array
      stopThemesCheckedAt:
        type: string
      topics:
        items:
          $ref: '#/definitions/models.Topic'
        type: array
      topicsCheckedAt:
        type: string
      withoutTopics:
        type: boolean
    type: object
  models.StopTheme:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
    type: object
  models.Topic:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
    type: object
  models.User:
    properties:
      channels:
        items:
          $ref: '#/definitions/models.Channel'
        type: array
      id:
        type: integer
      stopThemes:
        items:
          $ref: '#/definitions/models.StopTheme'
        type: array
      topics:
        items:
          $ref: '#/definitions/models.Topic'
        type: array
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for SmartFlow application
  title: SmartFlow REST API
  version: "1.0"
paths:
...
  /users/{id}:
    delete:
      description: Removes user and cleans their many-to-many associations
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      description: Retrieves user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates username and/or password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user
      tags:
      - users
...
swagger: "2.0"